+++ R/plot_heat.R
@@ -11,32 +11,32 @@
 #' @return
 #' **PlotHeat()** returns a heat plot.
 #'
-#' @seealso \code{\link{GetMetrics}} \code{\link{SetMetrics}}
-#'          \code{\link{PlotBeads}} \code{\link{PlotBar}}
-#'          \code{\link{PlotLine}}
+#' @seealso \code{\link{GetMetrics}}, \code{\link{SetMetrics}},
+#'          \code{\link{PlotBeads}}, \code{\link{PlotBar}},
+#'          \code{\link{PlotLine}}, \code{\link{PlotSpie}}
 #'
 #' @examples
 #' ## Not run:
-#' library(netmeta)
-#' data(Senn2013)
-#' nma <- netmeta(TE, seTE, treat1, treat2,
-#'   studlab, data = Senn2013, sm = "SMD")
+#' #library(netmeta)
+#' #data(Senn2013)
+#' #nma <- netmeta(TE, seTE, treat1, treat2,
+#' #studlab, data = Senn2013, sm = "SMD")
 #'
 #' # Get SUCRA
-#' nma.1 <- GetMetrics(nma, outcome = "HbA1c.random", prefer = "small", metrics = "SUCRA",
-#'   model = "random", simt = 1000)
-#' nma.2 <- GetMetrics(nma, outcome = "HbA1c.common", prefer = "small", metrics = "SUCRA",
-#'   model = "common", simt = 1000)
+#' #nma.1 <- GetMetrics(nma, outcome = "HbA1c.random", prefer = "small", metrics = "SUCRA",
+#' #model = "random", simt = 1000)
+#' #nma.2 <- GetMetrics(nma, outcome = "HbA1c.common", prefer = "small", metrics = "SUCRA",
+#' #model = "common", simt = 1000)
 #'
 #' # Combine metrics of multiple outcomes
-#' dataMetrics <- rbind(nma.1, nma.2)
+#' #dataMetrics <- rbind(nma.1, nma.2)
 #'
 #' # Set data for rankinma
-#' dataRankinma <- SetMetrics(dataMetrics, tx = tx, outcome = outcome,
-#'   metrics = SUCRA, metrics.name = "SUCRA")
+#' #dataRankinma <- SetMetrics(dataMetrics, tx = tx, outcome = outcome,
+#' #metrics = SUCRA, metrics.name = "SUCRA")
 #'
 #' # Illustrate heat plot
-#' PlotHeat(data = dataRankinma)
+#' #PlotHeat(data = dataRankinma)
 #' ## End(Not run)
 #'
 #' @export PlotHeat
@@ -58,9 +58,9 @@ PlotHeat <- function(data,
                       FALSE)
 
   lgcSorttx   <- ifelse(is.null(sorttx), FALSE,
-                      ifelse(argSorttx == TRUE, FALSE,
-                             ifelse(argSorttx == FALSE, FALSE,
-                                    TRUE)))
+                        ifelse(argSorttx == TRUE, FALSE,
+                               ifelse(argSorttx == FALSE, FALSE,
+                                      TRUE)))
 
   if (lgcInher) {
     cat(paste(" Inherit -------------------------------------------------- X\n",
@@ -105,7 +105,7 @@ PlotHeat <- function(data,
     aggregate(dataHeat$txs, by = list(dataHeat$outcome), "length"),
     aggregate(dataHeat$outcomes, by = list(dataHeat$outcome), "unique"),
     aggregate(dataHeat$tx, by = list(dataHeat$outcome), "paste")
-    )
+  )
 
 
   dataSmry$name.tx <- NA